+++ R/diff.R
@@ -1,4 +1,4 @@
-# Copyright (C) 2017  Brodie Gaslam
+# Copyright (C) 2018  Brodie Gaslam
 #
 # This file is part of "diffobj - Diffs for R Objects"
 #
@@ -190,7 +190,9 @@ make_diff_fun <- function(capt_fun) {
 #'       sequences
 #'     \item \dQuote{ansi256}: like \dQuote{ansi8}, except using the full range
 #'       of ANSI formatting options
-#'     \item \dQuote{html}: color and format using HTML markup
+#'     \item \dQuote{html}: color and format using HTML markup; the resulting
+#'       string is processed with \code{\link{enc2utf8}} when output as a full
+#'       web page (see docs for \code{html.output} under \code{\link{Style}}).
 #'   }
 #'   Defaults to \dQuote{auto}.  See \code{palette.of.styles} for details
 #'   on customization, \code{\link{style}} for full control of output format.